Upgrade to structopt 0.3 and fix resulting clippy complaints

+++ template/src/helpers.rs
@@ -1,16 +1,19 @@
 /*! Functions and templates which can be imported by app.rs to save effort */
 // Copyright 2017-2019, Stephan Sokolow
 
+// FIXME: Report that StructOpt is tripping Clippy's `result_unwrap_used` lint (which I use to push
+// for .expect() instead) in my two Option<T> fields and the `allow` gets ignored unless I
+// `#![...]` it onto the entire module.
+#![allow(clippy::result_unwrap_used)]
+
 use structopt::{clap, StructOpt};
 
 /// Modified version of Clap's default template for proper help2man compatibility
 ///
 /// Used as a workaround for:
-/// 1. StructOpt making display of `author` opt-out and not inheriting your preference in
-///    subcommands. ([TeXitoi/structopt/#172](https://github.com/TeXitoi/structopt/issues/172))
-/// 2. Clap's default template interfering with `help2man`'s proper function
+/// 1. Clap's default template interfering with `help2man`'s proper function
 ///    ([clap-rs/clap/#1432](https://github.com/clap-rs/clap/issues/1432))
-/// 3. Workarounds involving injecting `\n` into the description breaking help output if used
+/// 2. Workarounds involving injecting `\n` into the description breaking help output if used
 ///    on subcommand descriptions.
 pub const HELP_TEMPLATE: &str = "{bin} {version}
 
@@ -27,16 +30,20 @@ USAGE:
 #[derive(StructOpt, Debug)]
 #[structopt(rename_all = "kebab-case")]
 pub struct BoilerplateOpts {
+
     // -- Arguments used by main.rs --
     // TODO: Move these into a struct of their own in something like helpers.rs
 
+    // FIXME: Report that StructOpt trips Clippy's `cast_possible_truncation` lint unless I use
+    //        `u64` for my `from_occurrences` inputs, which is a ridiculous state of things.
+
     /// Decrease verbosity (-q, -qq, -qqq, etc.)
     #[structopt(short, long, parse(from_occurrences))]
-    pub quiet: usize,
+    pub quiet: u64,
 
     /// Increase verbosity (-v, -vv, -vvv, etc.)
     #[structopt(short, long, parse(from_occurrences))]
-    pub verbose: usize,
+    pub verbose: u64,
 
     /// Display timestamps on log messages (sec, ms, ns, none)
     #[structopt(short, long, value_name = "resolution")]
